backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose
pytest
pytest-asyncio
httpx